customer_order_service/requirements.txt
africastalking==1.2.8
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
djangorestframework==3.15.2
idna==3.9
josepy==1.14.0
mozilla-django-oidc==4.0.1
psycopg2==2.9.9
pycparser==2.22
pyOpenSSL==24.2.1
requests==2.32.3
schema==0.7.7
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
authlib == 1.0
django == 4.0
python-dotenv == 0.19
requests == 2.27